+++ src/components/content/diagnostic-reports/helpers/columns.tsx
@@ -1,3 +1,5 @@
+import { faArrowTrendUp } from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { ResourceTitleColumn } from "../../resource/helpers/resource-title-column";
 import { TableColumn } from "@/components/core/table/table-helpers";
 import { DiagnosticReportModel } from "@/fhir/models";
@@ -13,6 +15,7 @@ export const patientDiagnosticReportsColumns = (
       <ResourceTitleColumn
         title={diagnostic.displayName}
         ownedByBuilder={diagnostic.ownedByBuilder(builderId)}
+        capitalizeTitle={false}
       />
     ),
   },
@@ -23,9 +26,17 @@ export const patientDiagnosticReportsColumns = (
     minWidth: 128,
   },
   {
+    widthPercent: 1,
+    minWidth: 32,
+    className: "ctw-trend-indicator-cell",
+    render: (diagnostic) =>
+      diagnostic.hasTrends ? <FontAwesomeIcon icon={faArrowTrendUp} /> : null,
+  },
+  {
     title: "Details",
     dataIndex: "details",
-    widthPercent: 40,
+    widthPercent: 39,
     minWidth: 200,
+    className: "ctw-details-cell",
   },
 ];