+++ src/fhir/diagnostic-report.ts
@@ -1,17 +1,20 @@
 import { useIncludeBasics } from "./basic";
+import { LOINC_ANALYTES, ObservationModel } from "./models/observation";
+import { usePatientObservationsTrendData } from "./observations";
 import { searchBuilderRecords, searchCommonRecords } from "./search-helpers";
 import { SYSTEM_ZUS_THIRD_PARTY } from "./system-urls";
 import { applyDiagnosticReportFilters } from "@/components/content/diagnostic-reports/helpers/diagnostic-report-query-filters";
 import { CTWRequestContext } from "@/components/core/providers/ctw-context";
 import { useFeatureFlaggedQueryWithPatient } from "@/components/core/providers/patient-provider";
 import { getIncludedResources } from "@/fhir/bundle";
-import { PatientModel } from "@/fhir/models";
+import { DiagnosticReportModel, PatientModel } from "@/fhir/models";
 import { useFQSFeatureToggle } from "@/hooks/use-feature-toggle";
 import { createGraphqlClient, fqsRequest } from "@/services/fqs/client";
 import {
   DiagnosticReportGraphqlResponse,
   diagnosticReportQuery,
 } from "@/services/fqs/queries/diagnostic-reports-query";
+import { keys } from "@/utils/nodash";
 import {
   QUERY_KEY_OTHER_PROVIDER_DIAGNOSTIC_REPORTS,
   QUERY_KEY_PATIENT_DIAGNOSTIC_REPORTS,
@@ -21,32 +24,34 @@ import { Telemetry } from "@/utils/telemetry";
 type SearchType = "builder" | "all";
 
 export function usePatientBuilderDiagnosticReports() {
+  const { data } = usePatientObservationsTrendData(keys(LOINC_ANALYTES));
   return useFeatureFlaggedQueryWithPatient(
     QUERY_KEY_PATIENT_DIAGNOSTIC_REPORTS,
-    [],
+    [data?.map((o) => o.id)], // Only use the IDs in our key (fixes issue with ciruclar references).
     "diagnosticReports",
     "req.timing.builder_diagnostic_reports",
-    diagnosticReportsFetcherFQS("builder"),
-    diagnosticReportsFetcherODS("builder")
+    diagnosticReportsFetcherFQS("builder", data ?? []),
+    diagnosticReportsFetcherODS("builder", data ?? [])
   );
 }
 
 export function usePatientAllDiagnosticReports() {
   const fqs = useFQSFeatureToggle("diagnosticReports");
+  const { data } = usePatientObservationsTrendData(keys(LOINC_ANALYTES));
 
   const query = useFeatureFlaggedQueryWithPatient(
     QUERY_KEY_OTHER_PROVIDER_DIAGNOSTIC_REPORTS,
-    [],
+    [data?.map((o) => o.id)], // Only use the IDs in our key (fixes issue with ciruclar references).
     "diagnosticReports",
     "req.timing.all_diagnostic_reports",
-    diagnosticReportsFetcherFQS("all"),
-    diagnosticReportsFetcherODS("all")
+    diagnosticReportsFetcherFQS("all", data ?? []),
+    diagnosticReportsFetcherODS("all", data ?? [])
   );
 
   return useIncludeBasics(query, fqs);
 }
 
-function diagnosticReportsFetcherODS(searchType: SearchType) {
+function diagnosticReportsFetcherODS(searchType: SearchType, trendData: ObservationModel[]) {
   const fetchFunction = searchType === "builder" ? searchBuilderRecords : searchCommonRecords;
   return async (requestContext: CTWRequestContext, patient: PatientModel) => {
     try {
@@ -58,7 +63,7 @@ function diagnosticReportsFetcherODS(searchType: SearchType) {
         Telemetry.countMetric(`req.count.${searchType}_diagnostic_reports.none`);
       }
       Telemetry.histogramMetric(`req.count.${searchType}_diagnostic_reports`, resources.length);
-      return applyDiagnosticReportFilters(resources, getIncludedResources(bundle));
+      return applyDiagnosticReportFilters(resources, getIncludedResources(bundle), trendData);
     } catch (e) {
       throw Telemetry.logError(
         e as Error,
@@ -68,7 +73,7 @@ function diagnosticReportsFetcherODS(searchType: SearchType) {
   };
 }
 
-function diagnosticReportsFetcherFQS(searchType: SearchType) {
+function diagnosticReportsFetcherFQS(searchType: SearchType, trendData: ObservationModel[]) {
   const fetchFunction =
     searchType === "builder" ? diagnosticReportBuilderQueryFQS : diagnosticReportCommonQueryFQS;
   return async (requestContext: CTWRequestContext, patient: PatientModel) => {
@@ -83,7 +88,24 @@ function diagnosticReportsFetcherFQS(searchType: SearchType) {
         "fqs",
       ]);
 
-      return applyDiagnosticReportFilters(result);
+      // Enrich the supplied trend data with its parent diagnostic report.
+      const enrichedTrendData = trendData.map((o) => {
+        const observation = o;
+        const diagnosticReport = result.find((dr) =>
+          dr.result?.some((r) => r.reference === `Observation/${observation.id}`)
+        );
+        if (diagnosticReport) {
+          observation.diagnosticReport = new DiagnosticReportModel(
+            diagnosticReport,
+            undefined,
+            undefined,
+            trendData
+          );
+        }
+        return observation;
+      });
+
+      return applyDiagnosticReportFilters(result, undefined, enrichedTrendData);
     } catch (e) {
       throw Telemetry.logError(
         e as Error,